Fix main jar path in *.cfg produced by jpackage
A file binaries/main/app/AppName/lib/app/AppName.cfg is produced and
before this change, the main jar was specified with an absolute path and
passed as `--main-jar` to jpackage, while it is also among the jar files
passed to it via `--input` where relative paths are dervied for the
config file. The result is that the main jar is effectively
specified twice, with the first occurrence not being resolvable at
runtime later. This does not neccessarily lead to problems, but it
could. Either way, this is a bug that should be fixed. When specified
correctly, jpackage notices that the main jar is already specified in
the beginning as the first library on the classpath and leaves out the
second instance from the input directory.