Use MajorUpgrade feature for MSI builds
As explained in the official HowTo at
https://wixtoolset.org/docs/v3/howtos/updates/major_upgrade/
it is necessary to add a <MajorUpgrade> to the Product in order for
upgrades to remove older versions of the installed product during
installation and to show an error on attempting a downgrade.

It looks like MajorUpgrade is a mechanism that replaces the older one
with <Upgrade> and <UpgradeVersion> which JPackage is using.